+++ src/components/super/i-static-page/i-static-page.ts
@@ -21,12 +21,14 @@ import { instanceCache } from 'core/data';
 import { setLocale, locale } from 'core/i18n';
 
 import type { AppliedRoute, InitialRoute } from 'core/router';
-import * as cookie from 'core/kv-storage/engines/cookie';
-import { webEngineFactory } from 'core/system-theme-extractor/engines/web';
+
+import * as cookies from 'core/cookies';
+import CookieStorage from 'core/kv-storage/engines/cookie';
+
+import { SystemThemeExtractorWeb } from 'components/super/i-static-page/modules/theme';
 
 import {
 
-	remoteState,
 	resetComponents,
 
 	GlobalEnvironment,
@@ -88,7 +90,6 @@ export default abstract class iStaticPage extends iPage {
 	 * A module for manipulating page metadata, such as the page title or description
 	 */
 	@system<iStaticPage>((o) => SSR ? new SSRPageMetaData(o) : new PageMetaData())
-
 	readonly pageMetaData!: PageMetaData;
 
 	/**
@@ -102,8 +103,14 @@ export default abstract class iStaticPage extends iPage {
 	 */
 	@system<iStaticPage>((o) => themeManagerFactory(
 		o,
-		cookie.syncLocalStorage,
-		webEngineFactory(o)
+		{
+			themeStorageEngine: new CookieStorage('v4ls', {
+				cookies: cookies.from(o.remoteState.cookies ?? document),
+				maxAge: 2 ** 31 - 1
+			}),
+
+			systemThemeExtractor: new SystemThemeExtractorWeb(o)
+		}
 	))
 
 	readonly theme: CanUndef<ThemeManager>;
@@ -130,7 +137,7 @@ export default abstract class iStaticPage extends iPage {
 	 * Initial value for the active route.
 	 * This field is typically used in cases of SSR and hydration.
 	 */
-	@system(() => remoteState.route)
+	@system((o) => o.remoteState.route)
 	initialRoute?: InitialRoute | this['CurrentPage'];
 
 	/**
@@ -140,7 +147,7 @@ export default abstract class iStaticPage extends iPage {
 	 */
 	@system<iStaticPage>({
 		atom: true,
-		init: (o) => o.initGlobalEnv(remoteState)
+		init: (o) => o.initGlobalEnv(o.remoteState)
 	})
 
 	globalEnv!: GlobalEnvironment;