+++ src/core/init/CHANGELOG.md
@@ -9,11 +9,17 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.58 (2024-02-14)
+
+#### :bug: Bug Fix
+
+* Fixed a memory leak
+
 ## v4.0.0-beta.57 (2024-02-13)
 
 #### :bug: Bug Fix
 
-* Fixed a memory leak in `core/init/semaphore`
+* Fixed a memory leak
 
 ## v4.0.0-beta.48 (2024-01-17)